Implement DirectHConvolution.

+++ tex/cconv.cc
@@ -72,7 +72,7 @@ int main(int argc, char* argv[])
   cout << "min padded buffer=" << n << endl;
   unsigned int log2n;
   // Choose next power of 2 for maximal efficiency.
-  for(log2n=0; n > (1 << log2n); log2n++);
+  for(log2n=0; n > ((unsigned int) 1 << log2n); log2n++);
   n=1 << log2n;
   cout << "n=" << n << endl;
   cout << "m=" << m << endl;
@@ -90,14 +90,14 @@ int main(int argc, char* argv[])
 
   double offset=0.0;
   seconds();
-  for(int i=0; i < N; ++i) {
+  for(unsigned int i=0; i < N; ++i) {
     seconds();
     offset += seconds();
   }
 
   if(false) 
     {
-    int L=m;
+    unsigned int L=m;
     Complex *f=FFTWComplex(2*L);
     for(unsigned int i=0; i < L; i++) f[2*i]=d[i];
     for(unsigned int i=0; i < L; i++) f[2*i+1]=-d[i];
@@ -120,7 +120,7 @@ int main(int argc, char* argv[])
     Complex *u=FFTWComplex(m);
     Complex *v=FFTWComplex(m);
     ImplicitConvolution C(m,u,v);
-    for(int i=0; i < N; ++i) {
+    for(unsigned int i=0; i < N; ++i) {
       init(f,g);
       seconds();
       C.convolve(f,g,u,v);
@@ -144,7 +144,7 @@ int main(int argc, char* argv[])
   if(pad) {
     sum=0.0;
     ExplicitConvolution C(n,m,f);
-    for(int i=0; i < N; ++i) {
+    for(unsigned int i=0; i < N; ++i) {
       // FFTW out-of-place cr routines destroy the input arrays.
       init(f,g);
       seconds();
